refactor: Move TrialStore into own module
